requirements.txt
adafruit-io
python-dotenv
raspberry-gpio-python